bug-1866796: move shared system test setup functions to a utils script
